Moving everything to Session (cannot start session).
