+++ joint-sample-streaming.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:/Users/yura/Xcode Projects/personal/JointSample/joint-sample-streaming.xcodeproj">
+      location = "self:">
    </FileRef>
 </Workspace>
